requirements.txt
flask
flask-sqlalchemy
flask-marshmallow 
marshmallow-sqlalchemy
psycopg2